Explicitly cast to uint64_t

+++ include/picongpu/plugins/IsaacPlugin.x.cpp
@@ -779,10 +779,14 @@ namespace picongpu
                     else
                     {
                         uint64_t const localNrOfCells
-                            = cellDescription->getGridLayout().sizeWithoutGuardND().productOfComponents();
-                        cellCount = localNrOfCells * numProc;
-                        particleCount = localNrOfCells * TYPICAL_PARTICLES_PER_CELL
-                                        * (pmacc::mp_size<VectorAllSpecies>::type::value) * numProc;
+                            = static_cast<uint64_t>(
+                                cellDescription->getGridLayout().sizeWithoutGuardND().productOfComponents()
+                            );
+                        cellCount = localNrOfCells * static_cast<uint64_t>(numProc);
+                        particleCount = localNrOfCells * static_cast<uint64_t>(
+                                          TYPICAL_PARTICLES_PER_CELL
+                                        * (pmacc::mp_size<VectorAllSpecies>::type::value)
+                                        * numProc);
                         lastNotify = getTicksUs();
                         if(rank == 0)
                         {
